Alter threading methodology
So the only reason we're using threading is so that the 'side-effect' functionality of the report
doesn't add time between invocation and delivery of the report.  The process may now finalize without
finishing; as in the report will be delivered but the extra data may not necessarily be finished writing.